refactor(deps): relocate index provisioning + DDL to internal/indexer (decouple PR-B' slice B2)
Slice B2 of relocating the monitor/supervisor's package-main deps to internal/.
Moves the index schema + provisioning out of cmd/bintrail (package main) so the
future cmd/bintrail-console supervisor can provision a per-source index DB.

Moved init.go → internal/indexer (new schema.go + provision.go):
- 9 DDL constants (ddlSchemaSnapshots/StreamState/IndexState/ArchiveState/
  TableFlags/Profiles/AccessRules/FKConstraints/SchemaChanges) — kept PRIVATE
  (grep-checked: only createIndexTables + their unit tests referenced them).
- buildPartitionDefs, buildBinlogEventsDDL — kept PRIVATE.
- createBinlogEventsTable → EXPORT CreateBinlogEventsTable (cmd integration
  tests call it directly).
- createIndexTables → EXPORT CreateIndexTables.
- ensureDatabase → EXPORT EnsureDatabase.

NOT a pure relocation — one deliberate behavior change: EnsureDatabase gains a
`log func(string)` param (same nil→no-op idiom as CreateIndexTables's logTable).
The old body always did `fmt.Printf("Database %q ready")`; now the caller decides.
`bintrail init` passes a stdout closure (output preserved verbatim, incl. its
pre-existing unconditional print). The monitor supervisor passes nil → the daemon
provisions SILENTLY (intended — no stray stdout from a long-running process).

The cmd-local `ddlBintrailServers`/`ddlBintrailServerChanges` var aliases are
DELETED; the moved CreateIndexTables references serverid.DDLBintrailServers/
DDLBintrailServerChanges directly (new indexer→serverid edge, cycle-free:
serverid is a leaf). testutil.InitIndexTables keeps its own forked DDL by design
(separate package, can't reference the consts; unifying = test-behavior risk).

Callers rewired: init.go (runInit), monitor.go (supervisor Start), and the cmd
integration tests (CreateBinlogEventsTable + EnsureDatabase(...,nil) + the
bintrail_servers DDL → serverid.*). 13 tests moved init_test.go → indexer
schema_test.go (buildPartitionDefs/buildBinlogEventsDDL/DDL-content — verbatim,
bare names since they're now in-package); cmd init_test.go keeps its cobra/
parseS3ARN/s3Instructions/runInit tests.

Verified: go build/vet clean (unit + integration); full unit suite green (incl.
the 13 moved indexer tests); supervisor integration (indexer.EnsureDatabase +
CreateIndexTables via monitorSupervisor.Start) + cmd provisioning integration +
the full E2E pipeline (real `bintrail init`) all pass against MySQL.
